move randomChoice to proper utils file and add more unit tests
