+++ src/utils.js
@@ -1015,3 +1015,33 @@ export const getSalePriceMultiplier = (completedAchievements = {}) => {
 
   return salePriceMultiplier
 }
+
+/**
+ * @param {Array} an array of objects each containing a `weight` property
+ * @returns {Object} one of the items from weightedOptions
+ */
+export function randomChoice(weightedOptions) {
+  let totalWeight = 0
+  let sortedOptions = []
+
+  for (let option of weightedOptions) {
+    totalWeight += option.weight
+    sortedOptions.push(option)
+  }
+
+  sortedOptions.sort(o => o.weight)
+
+  let diceRoll = Math.random() * totalWeight
+  let option
+  let runningTotal = 0
+
+  for (let i in sortedOptions) {
+    option = sortedOptions[i]
+
+    if (diceRoll < option.weight + runningTotal) {
+      return option
+    }
+
+    runningTotal += option.weight
+  }
+}
